+++ R/sanitize_type.R
@@ -50,6 +50,7 @@ sanitize_type = function(settings) {
     "spline",
     "summary",
     "text",
+    "tile", "heatmap",
     "violin",
     "vline"
   )
@@ -114,6 +115,8 @@ sanitize_type = function(settings) {
       "spline"     = type_spline,
       "summary"    = type_summary,
       "text"       = type_text,
+      "tile"       = type_tile,
+      "heatmap"    = type_heatmap,
       "violin"     = type_violin,
       "vline"      = type_vline,
       type # default case (incl. line-family chars, handled below)